refactor: 기존 팝업 UI entity로 이동
