+++ src/entities/popup/api/index.ts
@@ -1,0 +1,1 @@
+export { type GetPopupApiResponse, getPopup } from './popup.api';
