+++ src/entities/popup/index.ts
@@ -1,0 +1,11 @@
+// api
+export { type GetPopupApiResponse, getPopup } from './api';
+
+// core
+export { type PopupRepository } from './core';
+
+// type
+export { type Popup } from './types';
+
+// ui
+export { PopupModal } from './ui';
