+++ src/entities/popup/ui/index.ts
@@ -1,0 +1,1 @@
+export { PopupModal } from './PopupModal';
